Add optional Serde support
When `serde` feature is explicitly enabled, we implement
`serde::Serialize` and `serde::Deserialize` for the `FlagsSet<T>`.

We don't implement Serialize+Deserialize for flags enum however, since
`FlagSet<T>` is what should be used anyway for flags and if users really
want to, they can easily use `serde_repr` crate to achieve that. We
document this easy method to make it even easier.